--4.22

+++ part4/backend/controllers/blogs.js
@@ -1,6 +1,6 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-const User = require('../models/user')
+const userExtractor = require('../utils/middleware').userExtractor
 
 blogsRouter.get('/', async (request, response, next) => {
   try {
@@ -19,18 +19,13 @@ blogsRouter.get('/:id', (request, response, next) => {
   }).catch(error => { next(error) })
 })
 
-blogsRouter.post('/', async (request, response, next) => {
+blogsRouter.post('/', userExtractor, async (request, response, next) => {
   try {
     const { title, url, author, likes } = request.body
     if (!title || !url) {
       return response.status(400).json({ error: 'title or url missing' })
     }
-
-    const user = await User.findById(request.user.id)
-    if (!user) {
-      return response.status(400).json({ error: 'no user found' })
-    }
-
+    const user = request.user
     const blog = new Blog({
       title,
       url,
@@ -38,7 +33,6 @@ blogsRouter.post('/', async (request, response, next) => {
       likes: likes || 0,
       user: user._id
     })
-
     const savedBlog = await blog.save()
     user.blogs = user.blogs.concat(savedBlog._id)
     await user.save()
@@ -73,14 +67,18 @@ blogsRouter.put('/:id', async (request, response, next) => {
   } catch (error) { next(error) }
 })
 
-blogsRouter.delete('/:id', async (request, response, next) => {
+blogsRouter.delete('/:id', userExtractor, async (request, response, next) => {
   try {
-    const result = await Blog.findByIdAndDelete(request.params.id)
-    if (result) {
-      response.status(204).end()
-    } else {
-      response.status(404).json({ error: 'blog not found' })
+    const blog = await Blog.findById(request.params.id)
+    if (!blog) {
+      return response.status(404).json({ error: 'blog not found' })
     }
+
+    if (blog.user.toString() !== request.user.id.toString()) {
+      return response.status(401).json({ error: 'not authorized' })
+    }
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
   } catch (error) { next(error) }
 })
